Fix makefile for linux.
activate_venv used a different venv location by platform, but the
Makefile did not. I have changed activate_venv to match the location in
.gitignore and the Makefile.

The shebang at the top of a Makefile does not seem to work for me on
Ubuntu 22 (actually Pop! OS), so I could not source the activate_venv
which uses bash syntax, instead of sh. Adding SHELL to the Makefile has
fixed this for me. So `make venv` no longer installs everything to the
user's global site-packages :)